requirements.txt
pandas
faiss-cpu
sentence-transformers
transformers
